// src/app/models/vehicle/vehicle.ts
export interface VehicleInterface {
  user_id: number;
  brand: string;
  model: string;
  color: string;
  licensePlate: string;
  mileage: string;
  year: any;
  clientName: string;
  clientPhoneNumber: string;
  observations: string;
  retrovisor_izquierdo: boolean;
  retrovisor_derecho: boolean;
  retrovisor_interno: boolean;
  alfombras: boolean;
  cepillos_de_parabrisas: boolean;
  antena: boolean;
  estereo: boolean;
  tapa_de_gasolina: boolean;
  tapa_valvula_de_aire: boolean;
  placa_delantera: boolean;
  placa_trasera: boolean;
  capo: boolean;
  compuerta_trasera: boolean;
  pintura: boolean;
  parabrisas: boolean;
  puertas_abrir_cerrar: boolean;
  manillas: boolean;
  cocuyos: boolean;
  emblema: boolean;
  parachoques: boolean;
  tapas_de_rines: boolean;
  cinturones_de_seguridad: boolean;
  neumaticos: boolean;
  techo: boolean;
  encendedor_de_cigarrillo: boolean;
  luces_neblina_delanteras: boolean;
  luces_traseras: boolean;
  luces_delanteras: boolean;
  luces_altas: boolean;
  corneta: boolean;
  alarma: boolean;
  vidrios_suben_bajan: boolean;
  luces_stock: boolean;
  luces_internas: boolean;
  aire_acondicionado: boolean;
  frenos: boolean;
  bateria: boolean;
  extinguidor: boolean;
  cables_para_corriente: boolean;
  gato: boolean;
  triangulo: boolean;
  llave_de_seguridad: boolean;
  llave_ajustar_tuercas: boolean;
  neumatico_repuesto: boolean;
  taza_de_ruedas: boolean;
  fuel_level: boolean;
}